--- src/Home.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-function Home() {
-  return <h1>ホーム画面</h1>;
-}
-export default Home;
