+++ src/main.jsx
@@ -1,10 +1,6 @@
-import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
-import "./App.css";
-import "./firebase.js" 
+import { BrowserRouter } from "react-router";
 import Router from "./routes/route.jsx";
-
 
 const root = createRoot(document.getElementById("root"));
