--- src/pages/Notification.jsx
@@ -1,7 +1,0 @@
-export default function Notification() {
-  return (
-    <div>
-      <h1>Notification</h1>
-    </div>
-  );
-}
