--- src/pages/Search.jsx
@@ -1,7 +1,0 @@
-export default function Search() {
-  return (
-    <div>
-      <h1>Search</h1>
-    </div>
-  );
-}
