--- src/pages/Timeline.jsx
@@ -1,7 +1,0 @@
-export default function Timeline() {
-  return (
-    <div>
-      <h1>Timeline</h1>
-    </div>
-  );
-}
