--- src/pages/User.jsx
@@ -1,7 +1,0 @@
-export default function User() {
-  return (
-    <div>
-      <h1>ユーザー情報</h1>
-    </div>
-  );
-}
